Feature/authentication2

+++ PickupPointBackEnd/src/main/java/com/tqs/pickuppointbackend/controller/PickupScheduleController.java
@@ -1,6 +1,8 @@
 package com.tqs.pickuppointbackend.controller;
 
 
+import com.tqs.pickuppointbackend.constants.UserType;
+import com.tqs.pickuppointbackend.service.AuthenticationService;
 import lombok.extern.log4j.Log4j2;
 
 import java.security.NoSuchAlgorithmException;
@@ -10,13 +12,7 @@ import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import com.tqs.pickuppointbackend.exceptions.ResourceNotFoundException;
 import com.tqs.pickuppointbackend.model.PickupSchedule;
@@ -30,73 +26,104 @@ public class PickupScheduleController {
     @Autowired
     PickupScheduleService pickupScheduleService;
 
-    @GetMapping("/pickupschedules")
-    public ResponseEntity<List<PickupSchedule>> getPickupSchedules(){
+    @Autowired
+    private AuthenticationService authenticationService;
 
-        return ResponseEntity.ok().body(pickupScheduleService.getPickupSchedules());
+
+    @GetMapping("/pickupschedules")
+    public ResponseEntity<List<PickupSchedule>> getPickupSchedules(@RequestHeader("Authorization") String token) throws ResourceNotFoundException {
+
+        if (authenticationService.hasAcess(token, UserType.ACP)){
+            return ResponseEntity.ok().body(pickupScheduleService.getPickupSchedules());
+        }
+        throw new ResourceNotFoundException("The user dont have acess");
 
     }
 
     @GetMapping("/pickupschedules/user/{id}")
-    public ResponseEntity<List<PickupSchedule>> getPickupSchedulesByuserId(@PathVariable(value = "id") long id) throws ResourceNotFoundException {
+    public ResponseEntity<List<PickupSchedule>> getPickupSchedulesByuserId(@PathVariable(value = "id") long id, @RequestHeader("Authorization") String token) throws ResourceNotFoundException {
 
-        return ResponseEntity.ok().body(pickupScheduleService.getPickupScheduleUserById(id));
+        if (authenticationService.hasAcess(token, UserType.CLIENT)){
+            return ResponseEntity.ok().body(pickupScheduleService.getPickupScheduleUserById(id));
+        }
+        throw new ResourceNotFoundException("The user dont have acess");
 
     }
 
     @GetMapping("/pickupschedule/{id}")
-    public ResponseEntity<PickupSchedule> getPickupScheduleById(@PathVariable(value = "id") long id) throws ResourceNotFoundException {
+    public ResponseEntity<PickupSchedule> getPickupScheduleById(@PathVariable(value = "id") long id,  @RequestHeader("Authorization") String token) throws ResourceNotFoundException {
 
-        return ResponseEntity.ok().body(pickupScheduleService.getPickupScheduleById(id));
+        if (authenticationService.hasAcess(token, UserType.ACP)){
+            return ResponseEntity.ok().body(pickupScheduleService.getPickupScheduleById(id));
+        }
+        throw new ResourceNotFoundException("The user dont have acess");
 
     }
 
     @PostMapping("/pickupschedule")
-    public ResponseEntity<PickupSchedule> addPickupSchedule(@RequestBody PickupScheduleDTO pickupScheduleDTO) throws ResourceNotFoundException, NoSuchAlgorithmException {
+    public ResponseEntity<PickupSchedule> addPickupSchedule(@RequestBody PickupScheduleDTO pickupScheduleDTO, @RequestHeader("Authorization") String token) throws ResourceNotFoundException, NoSuchAlgorithmException {
 
-        return ResponseEntity.ok().body(pickupScheduleService.addPickupSchedule(pickupScheduleDTO));
-    
+        if (authenticationService.hasAcess(token, UserType.CLIENT)){
+            return ResponseEntity.ok().body(pickupScheduleService.addPickupSchedule(pickupScheduleDTO));
+        }
+        throw new ResourceNotFoundException("The user dont have acess");
+
     }
 
     @GetMapping("/pickupschedules/available")
-    public ResponseEntity<List<PickupSchedule>> getPickupScheduleByAvailabilty() throws ResourceNotFoundException {
+    public ResponseEntity<List<PickupSchedule>> getPickupScheduleByAvailabilty(@RequestHeader("Authorization") String token) throws ResourceNotFoundException {
 
-        return ResponseEntity.ok().body(pickupScheduleService.getAvailablePickupSchedules());
+        if (authenticationService.hasAcess(token, UserType.ACP)){
+            return ResponseEntity.ok().body(pickupScheduleService.getAvailablePickupSchedules());
+        }
+        throw new ResourceNotFoundException("The user dont have acess");
 
     }
 
     @GetMapping("/pickupschedules/available/{id}")
-    public ResponseEntity<List<PickupSchedule>> getPickupScheduleByAvailabiltyByPickupPointId(@PathVariable(value = "id") long id) throws ResourceNotFoundException {
+    public ResponseEntity<List<PickupSchedule>> getPickupScheduleByAvailabiltyByPickupPointId(@PathVariable(value = "id") long id, @RequestHeader("Authorization") String token) throws ResourceNotFoundException {
 
-        return ResponseEntity.ok().body(pickupScheduleService.getAvailablePickupSchedulesByPickupPointId(id));
+        if (authenticationService.hasAcess(token, UserType.ACP)){
+            return ResponseEntity.ok().body(pickupScheduleService.getAvailablePickupSchedulesByPickupPointId(id));
+        }
+        throw new ResourceNotFoundException("The user dont have acess");
 
     }
 
     @GetMapping("/pickupschedules/nonavailable/{id}")
-    public ResponseEntity<List<PickupSchedule>> getPickupScheduleByNonAvailabiltyByPickupPointId(@PathVariable(value = "id") long id) throws ResourceNotFoundException {
+    public ResponseEntity<List<PickupSchedule>> getPickupScheduleByNonAvailabiltyByPickupPointId(@PathVariable(value = "id") long id, @RequestHeader("Authorization") String token) throws ResourceNotFoundException {
 
-        return ResponseEntity.ok().body(pickupScheduleService.getNonAvailablePickupSchedulesByPickupPointId(id));
+        if (authenticationService.hasAcess(token, UserType.ACP)){
+            return ResponseEntity.ok().body(pickupScheduleService.getNonAvailablePickupSchedulesByPickupPointId(id));
+        }
+        throw new ResourceNotFoundException("The user dont have acess");
 
     }
 
     @GetMapping("/pickupschedules/nonavailable")
-    public ResponseEntity<List<PickupSchedule>> getPickupScheduleByNonAvailabilty() throws ResourceNotFoundException {
-
-        return ResponseEntity.ok().body(pickupScheduleService.getNonAvailablePickupSchedules());
+    public ResponseEntity<List<PickupSchedule>> getPickupScheduleByNonAvailabilty(@RequestHeader("Authorization") String token) throws ResourceNotFoundException {
+        if (authenticationService.hasAcess(token, UserType.ACP)){
+            return ResponseEntity.ok().body(pickupScheduleService.getNonAvailablePickupSchedules());
+        }
+        throw new ResourceNotFoundException("The user dont have acess");
 
     }
 
     @PutMapping("/pickupschedule/{id}")
-    public ResponseEntity<PickupSchedule> updatePickupSchedule(@PathVariable(value = "id") long id, @Valid @RequestBody PickupScheduleDTO pickupScheduleDTO) throws ResourceNotFoundException {
-        System.out.println(id);
-        return ResponseEntity.ok(pickupScheduleService.updatePickupSchedule(pickupScheduleDTO));
+    public ResponseEntity<PickupSchedule> updatePickupSchedule(@PathVariable(value = "id") Long id, @Valid @RequestBody PickupScheduleDTO pickupScheduleDTO, @RequestHeader("Authorization") String token) throws ResourceNotFoundException {
+        if (authenticationService.hasAcess(token, UserType.ACP)){
+            return ResponseEntity.ok(pickupScheduleService.updatePickupSchedule(pickupScheduleDTO));
+        }
+        throw new ResourceNotFoundException("The user dont have acess");
     }
 
     @DeleteMapping("/pickupschedule/{id}")
-    public ResponseEntity<PickupSchedule> deletePickupScheduleById(@PathVariable(value = "id") long id) throws ResourceNotFoundException {
+    public ResponseEntity<PickupSchedule> deletePickupScheduleById(@PathVariable(value = "id") long id, @RequestHeader("Authorization") String token) throws ResourceNotFoundException {
+        if (authenticationService.hasAcess(token, UserType.ACP)){
+            return ResponseEntity.ok().body(pickupScheduleService.deletePickupScheduleById(id));
+        }
+        throw new ResourceNotFoundException("The user dont have acess");
 
-        return ResponseEntity.ok().body(pickupScheduleService.deletePickupScheduleById(id));
-    
     }
